Relocate models in their own subpackages
