--- src/main/java/by/draughts/model/Tournament.java
@@ -1,4 +1,0 @@
-package by.draughts.model;
-
-public class Tournament {
-}
