+++ src/main/java/by/draughts/model/game/GameResult.java
@@ -1,4 +1,4 @@
-package by.draughts.model;
+package by.draughts.model.game;
 
 public enum GameResult {
     WHITE_WON, DRAW, BLACK_WON, WHITE_WON_TIME, BLACK_WON_TIME,
