+++ src/main/java/by/draughts/model/ply/Ply.java
@@ -1,4 +1,6 @@
-package by.draughts.model;
+package by.draughts.model.ply;
+
+import by.draughts.model.game.Position;
 
 import java.util.List;
