+++ src/main/java/by/draughts/model/ply/PlyRate.java
@@ -1,4 +1,4 @@
-package by.draughts.model;
+package by.draughts.model.ply;
 
 public enum PlyRate {
     GOOD, EXCELLENT, BAD, TERRIBLE, DISPUTABLE, SHARP
