+++ src/main/java/by/draughts/model/tournament/Tournament.java
@@ -1,0 +1,4 @@
+package by.draughts.model.tournament;
+
+public class Tournament {
+}
