+++ src/main/java/by/draughts/service/impl/GameServiceImpl.java
@@ -1,8 +1,7 @@
 package by.draughts.service.impl;
 
-import by.draughts.model.Game;
+import by.draughts.model.game.Game;
 import by.draughts.service.GameService;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
